backend/auth-service/requirements.txt
fastapi
uvicorn
pyjwt
